refactor: extract keybindings overlay + nested profile picker (#357)
Ninth slice. Two related overlays bundled in one file:

- draw_keybindings                  (~123 lines) - main keybindings list with rebind capture
- draw_keybindings_profile_picker   (~70 lines)  - nested profile picker, kept private

draw_keybindings_profile_picker is only ever invoked from draw_keybindings
when keybindings_overlay.profile_picker is true, so it stays private to
the same submodule rather than getting its own file.

The keybindings module's `self` and BindingMode imports moved with the
extraction; KeyAction remains in ui/mod.rs because draw_help still uses it.

ui/mod.rs: 4132 -> 3938 lines.
